usb: dwc2: add support for Canaan K230
Canaan Kendryte K230 SoC instantiates a dwc2 4.30a core. This patch
adds the compatible for it.

Signed-off-by: Junhui Liu <liujh2818@outlook.com>
---
The USB function on K230 currently relies on Kongyang's patch [1], which
updated the core reset flow of dwc2 in version 4.20 and newer. However,
this patch was later included in another updated patch also from
Kongyang [2], but this patch lost the update and conflicts with the
master branch. I will take over the work [2] and update it later later.

[1] https://lore.kernel.org/u-boot/20240328131811.94559-1-seashell11234455@gmail.com/
[2] https://lore.kernel.org/u-boot/20240522142311.17351-1-seashell11234455@gmail.com/