riscv: dts: canaan: Add basic device tree for K230 CanMV board
Add initial dts for K230-CanMV powered by Canaan Kendryte K230 SoC,
which has two RISC-V C908 cores, a big core with vector 1.0 extension
and a small core without vector extension.

This patch is basically comes from Linux Kernel [1] and it assumes
u-boot is running on the big core. Additionally, boot_ctl and reboot
nodes are added to support sysreset and an clk_dummy node is added to
satisfy dependencies for usb.

Currently, u-boot is booted by the vendor's u-boot-spl. To meet the
requirements of vendor's u-boot-spl for u-boot, a binman node with
mkimage child node is added here, which will compress u-boot.bin with
gzip and generate an image named "uboot" in the file u-boot-gz.img.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/conor/linux.git/log/?h=k230-basic

Signed-off-by: Junhui Liu <junhui.liu@pigmoral.tech>